requirements.txt
pandas==1.5.0
loguru==0.6.0
parsel==1.6.0
mysql-connector-python==8.0.31
python-dotenv==0.21.0